requirements.txt
fastapi
pydantic
uvicorn
espnet
SoundFile
espnet_model_zoo
streamlit
